+++ core/page_iterator.go
@@ -2,6 +2,8 @@ package core
 
 import (
 	"net/url"
+
+	"github.com/michaeldcanady/servicenow-sdk-go/internal"
 )
 
 // Deprecated: deprecated in v1.5.0. Please use PageIterator2[T].
@@ -15,7 +17,7 @@ type PageIterator[T any, C CollectionResponse[T]] struct {
 // Deprecated: deprecated in v1.5.0. Please use NewPageIterator2[T].
 // NewPageIterator creates a new PageIterator instance.
 func NewPageIterator[T any, C CollectionResponse[T]](currentPage CollectionResponse[T], client Client) (*PageIterator[T, C], error) {
-	if isNil(client) {
+	if internal.IsNil(client) {
 		return nil, ErrNilClient
 	}