+++ core/request_builder.go
@@ -6,6 +6,7 @@ import (
 	"reflect"
 
 	"github.com/gabriel-vasile/mimetype"
+	"github.com/michaeldcanady/servicenow-sdk-go/internal"
 )
 
 // RequestBuilder represents a builder for constructing HTTP request information.
@@ -197,13 +198,13 @@ func (rB *RequestBuilder) ToRequestInformation3(method HttpMethod, config *Reque
 	requestInfo := NewRequestInformation()
 
 	if config != nil {
-		if !isNil(config.QueryParameters) {
+		if !internal.IsNil(config.QueryParameters) {
 			err := requestInfo.AddQueryParameters(config.QueryParameters)
 			if err != nil {
 				return nil, err
 			}
 		}
-		if !isNil(config.Data) {
+		if !internal.IsNil(config.Data) {
 			data, err := rB.prepareData(config.Data)
 			if err != nil {
 				return nil, err
@@ -269,11 +270,7 @@ func (rB *RequestBuilder) SendGet(params interface{}, errorMapping ErrorMapping,
 }
 
 func (rB *RequestBuilder) SendGet2(config *RequestConfiguration) error {
-	err := SendGet2(rB, config)
-	if err != nil {
-		return err
-	}
-	return nil
+	return rB.sendRequest(GET, config)
 }
 
 // Deprecated: deprecated since v1.4.0. Please use SendPost3
@@ -282,7 +279,7 @@ func (rB *RequestBuilder) SendPost(data map[string]string, params interface{}, e
 }
 
 func (rB *RequestBuilder) SendPost3(config *RequestConfiguration) error {
-	return SendPost2(rB, config)
+	return rB.sendRequest(POST, config)
 }
 
 // Deprecated: deprecated since v1.4.0. Please use SendPost3
@@ -296,7 +293,7 @@ func (rB *RequestBuilder) SendDelete(params interface{}, errorMapping ErrorMappi
 }
 
 func (rB *RequestBuilder) SendDelete2(config *RequestConfiguration) error {
-	return sendDelete2(rB, config)
+	return rB.sendRequest(DELETE, config)
 }
 
 // Deprecated: deprecated since v1.4.0. Please use SendPut2
@@ -305,5 +302,23 @@ func (rB *RequestBuilder) SendPut(data map[string]string, params interface{}, er
 }
 
 func (rB *RequestBuilder) SendPut2(config *RequestConfiguration) error {
-	return sendPut2(rB, config)
+	return rB.sendRequest(PUT, config)
+}
+
+func (rB *RequestBuilder) sendRequest(method HttpMethod, config *RequestConfiguration) error {
+	requestInfo, err := rB.ToRequestInformation3(method, config)
+	if err != nil {
+		return err
+	}
+
+	response, err := rB.Client.Send(requestInfo, config.ErrorMapping.(ErrorMapping))
+	if err != nil {
+		return err
+	}
+
+	if method != DELETE {
+		return ParseResponse(response, &config.Response)
+	}
+
+	return nil
 }