+++ table-api/table_item_request_builder2.go
@@ -1,25 +1,54 @@
 package tableapi
 
-import "github.com/michaeldcanady/servicenow-sdk-go/core"
+import (
+	"github.com/michaeldcanady/servicenow-sdk-go/core"
+	"github.com/michaeldcanady/servicenow-sdk-go/internal"
+	intTable "github.com/michaeldcanady/servicenow-sdk-go/table-api/internal"
+)
 
-// Deprecated: deprecated since v{unreleased}. Use `TableItemRequestBuilder2` instead.
-type TableItemRequestBuilder struct {
-	core.RequestBuilder
+const (
+	tableItemURLTemplate = "{+baseurl}/table{/table}{/sysId}{?sysparm_display_value,sysparm_exclude_reference_link,sysparm_fields,sysparm_input_display_value,sysparm_query_no_domain,sysparm_view,sysparm_query_no_domain}"
+)
+
+type TableItemRequestBuilder2 struct {
+	intTable.RequestBuilder
 }
 
-// Deprecated: deprecated since v{unreleased}. Use `NewTableItemRequestBuilder2` instead.
-// NewTableItemRequestBuilder creates a new instance of the TableItemRequestBuilder associated with the given URL and Client.
-// It accepts the URL and Client as parameters and returns a pointer to the created TableItemRequestBuilder.
-func NewTableItemRequestBuilder(client core.Client, pathParameters map[string]string) *TableItemRequestBuilder {
-	requestBuilder := core.NewRequestBuilder(
-		client,
-		"{+baseurl}/table{/table}{/sysId}{?sysparm_display_value,sysparm_exclude_reference_link,sysparm_fields,sysparm_input_display_value,sysparm_query_no_domain,sysparm_view,sysparm_query_no_domain}",
-		pathParameters,
-	)
-	return &TableItemRequestBuilder{
-		*requestBuilder,
+// NewTableItemRequestBuilder2 creates a new instance of TableItemRequestBuilder2.
+func NewTableItemRequestBuilder2(client core.Client, pathParameters map[string]string) (*TableItemRequestBuilder2, error) {
+	if internal.IsNil(client) {
+		return nil, ErrNilClient
 	}
+
+	_, basePathOk := pathParameters[internal.BasePathParameter]
+	if !basePathOk {
+		return nil, core.ErrMissingBasePathParam
+	}
+
+	_, tableOk := pathParameters["table"]
+	if !tableOk {
+		return nil, ErrNilParameterTable
+	}
+
+	_, sysIDOk := pathParameters["sysId"]
+	if !sysIDOk {
+		return nil, ErrNilParameterSysID
+	}
+
+	return &TableItemRequestBuilder2{
+		RequestBuilder: core.NewRequestBuilder(client, tableItemURLTemplate, pathParameters),
+	}, nil
 }
+
+// TableItemService provides an interface for the service methods, adhering to the Interface Segregation Principle.
+type TableItemService interface {
+	Get(params *TableItemRequestBuilderGetQueryParameters) (*TableItemResponse, error)
+	Delete(params *TableItemRequestBuilderDeleteQueryParameters) error
+	Put(tableEntry interface{}, params *TableItemRequestBuilderPutQueryParameters) (*TableItemResponse, error)
+}
+
+// Ensure that TableItemRequestBuilder implements TableItemService.
+var _ TableItemService = (*TableItemRequestBuilder2)(nil)
 
 // Get sends an HTTP GET request using the specified query parameters and returns a TableItemResponse.
 //
@@ -29,7 +58,7 @@ func NewTableItemRequestBuilder(client core.Client, pathParameters map[string]st
 // Returns:
 //   - *TableItemResponse: The response data as a TableItemResponse.
 //   - error: An error if there was an issue with the request or response.
-func (rB *TableItemRequestBuilder) Get(params *TableItemRequestBuilderGetQueryParameters) (*TableItemResponse, error) {
+func (rB *TableItemRequestBuilder2) Get(params *TableItemRequestBuilderGetQueryParameters) (*TableItemResponse, error) {
 	config := &tableItemGetRequestConfiguration2[TableEntry]{
 		header:   nil,
 		query:    params,
@@ -52,7 +81,7 @@ func (rB *TableItemRequestBuilder) Get(params *TableItemRequestBuilderGetQueryPa
 //
 // Returns:
 //   - error: An error if there was an issue with the request or response, or nil if the request was successful.
-func (rB *TableItemRequestBuilder) Delete(params *TableItemRequestBuilderDeleteQueryParameters) error {
+func (rB *TableItemRequestBuilder2) Delete(params *TableItemRequestBuilderDeleteQueryParameters) error {
 	config := &tableItemDeleteRequestConfiguration2[TableEntry]{
 		header:   nil,
 		query:    params,
@@ -63,24 +92,7 @@ func (rB *TableItemRequestBuilder) Delete(params *TableItemRequestBuilderDeleteQ
 	return rB.SendDelete2(config.toConfiguration())
 }
 
-// Deprecated: deprecated since v1.4.0. Use `Put2` instead.
-//
 // Put updates a table item using an HTTP PUT request.
-// It takes a map of table entry data and optional query parameters to send in the request.
-// The method returns a TableItemResponse representing the updated item or an error if the request fails.
-//
-// Parameters:
-//   - tableEntry: A map containing the data to update the table item.
-//   - params: An optional pointer to TableItemRequestBuilderPutQueryParameters, which can be used to specify query parameters for the request.
-//
-// Returns:
-//   - *TableItemResponse: A TableItemResponse containing the updated item data.
-//   - error: An error, if the request fails at any point, such as request information creation or JSON deserialization.
-func (rB *TableItemRequestBuilder) Put(tableEntry map[string]string, params *TableItemRequestBuilderPutQueryParameters) (*TableItemResponse, error) {
-	return rB.Put2(tableEntry, params)
-}
-
-// Put2 updates a table item using an HTTP PUT request.
 // It takes a map of table entry data and optional query parameters to send in the request.
 // The method returns a TableItemResponse representing the updated item or an error if the request fails.
 //
@@ -91,7 +103,7 @@ func (rB *TableItemRequestBuilder) Put(tableEntry map[string]string, params *Tab
 // Returns:
 //   - *TableItemResponse: A TableItemResponse containing the updated item data.
 //   - error: An error, if the request fails at any point, such as request information creation or JSON deserialization.
-func (rB *TableItemRequestBuilder) Put2(tableEntry interface{}, params *TableItemRequestBuilderPutQueryParameters) (*TableItemResponse, error) {
+func (rB *TableItemRequestBuilder2) Put(tableEntry interface{}, params *TableItemRequestBuilderPutQueryParameters) (*TableItemResponse, error) {
 	tableEntry, err := convertFromTableEntry(tableEntry)
 	if err != nil {
 		return nil, err